requirements.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.0.1
pandas==2.1.3
openpyxl==3.1.2
numpy==1.25.2